pyproject.toml
[project]
name = "hestia-agent"
version = "0.1.0"
description = "Hestia personal assistant — FastAPI agent service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic-settings>=2.7",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "bcrypt>=4.0,<5.0",
    "anthropic>=0.49",
    "aiosqlite>=0.21",
    "python-multipart>=0.0.20",
    "httpx>=0.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
    "coverage>=7.0",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "T20", "SIM"]
ignore = ["S105", "S106"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]
